[Chore #163239568] Add new route file for comments
