+++ server/routes/index.js
@@ -4,6 +4,7 @@ import questionRouter from './question';
 import rsvpRouter from './rsvp';
 import userRouter from './user';
 import userAdminRouter from './admin-users';
+import commentRouter from './comment';
 import Auth from '../middlewares/auth';
 import Misc from '../middlewares/misc';
 
@@ -25,5 +26,6 @@ router.use('/api/v1', trimBody, checkToken, meetupRouter);
 router.use('/api/v1', trimBody, checkToken, rsvpRouter);
 router.use('/api/v1', trimBody, checkToken, questionRouter);
 router.use('/api/v1', trimBody, checkToken, userAdminRouter);
+router.use('/api/v1', trimBody, checkToken, commentRouter);
 
 export default router;
